Added combManif.pl which consolidates data from previous BackupManif files
Previously, sort.pl creates a text file that details all of the music that was
sorted during the runtime of the program. However, after a few runs, the
destination directory can get pretty cluttered. When it's time to back up the
music, this file creates a single file for all of the previous BackupManif's.

KNOWN BUGS: Characters are sorted by Unicode value. Unfortunately, this means
that artist names that start with a lowercase letter are not sorted correctly.
Patch coming soon.